requirements.txt
pyyaml
SQLAlchemy
psycopg2
pandas
requests
pyodbc
openpyxl
kafka-python

